Run linting with npm test
I've gone back and forth about this a bunch of times, but I've come to
think this is the best approach. It's true that linting isn't exactly
the same as testing, but most people just run "npm test" and I wouldn't
want them to miss out on our linting results.

The "validate" script remains as an alias of "test" in case anyone is
using that out of habit.