Fix #69: Provide multiple dataset endpoints
Before this commit, we served one giant JSON blob for each dataset. For
example, /datasets/user-activity included all user activity data for all
metrics in all categories.

With this change, each dataset now has multiple endpoints. Each endpoint
serves a portion of the data. For example, /datasets/user-activity lists
all category and metric names. /datasets/user-activity/Russia/YAU serves
only the YAU data in the Russia category.

This should allow us to improve the performance of the Ensemble
front-end. For example, this will allow us to fix the issue of the
Hardware dashboard taking about 5 seconds to load even on a good
connection.

I already have a commit to Ensemble with the necessary changes which
will be published soon.