Remove support for local data sources
Before the Firefox Public Data Report was launched, much of the data
that ensemble-transposer ingested was secret. To keep it secret, we
downloaded it locally and added support for local data sources.

All of the data that ensemble-transposer ingests is now public, so there
is no longer a need for this support. If we do need to work with secret
data in the future, we could just host it locally with the "serve" npm
module (or similar) to keep things simple.