Issue #81, #24, fix #64: Refactor for redash
Refactor this project to support a variety of different input formats.
This will help us transpose data from redash.

This is a major rewrite, so I also took the opportunity to improve GCP
support. We now write transposed data to local files on a regular basis.
With some relatively minor changes, this could be a Google Cloud
Function which writes to Google Cloud Storage.

For the time being, a server is provided to host the contents of the
generated files. We can remove the server once DataOps is ready for us
to migrate to GCP.